+++ plugins/tracker/src/index.ts
@@ -284,10 +284,12 @@ export default plugin(trackerId, {
     SetProject: '' as Ref<Action>,
     CopyIssueId: '' as Ref<Action>,
     CopyIssueTitle: '' as Ref<Action>,
+    CopyIssueLink: '' as Ref<Action>,
     MoveToTeam: '' as Ref<Action>,
     Relations: '' as Ref<Action>,
     CopyIssueLink: '' as Ref<Action>,
-    NewSubIssue: '' as Ref<Action>
+    NewSubIssue: '' as Ref<Action>,
+    EditWorkflowStatuses: '' as Ref<Action>
   },
   team: {
     DefaultTeam: '' as Ref<Team>