+++ changelog.md
@@ -5,9 +5,10 @@
 Core:
 
 - Allow to leave workspace
-- Allow to kick employee
+- Allow to kick employee (Only for owner)
 - Browser notifications
 - Allow to create employee
+- Owner role for employee
 
 HR:
 
@@ -16,6 +17,7 @@ HR:
 Tracker:
 - Manual issues ordering
 - Issue relations
+- Issue status management
 
 Workbench
 - Use application aliases in URL